fix delete_user_usecase.py and TU

+++ src/modules/delete_user/app/delete_user_usecase.py
@@ -1,5 +1,6 @@
 from src.shared.domain.entities.user import User
 from src.shared.domain.repositories.user_repository_interface import IUserRepository
+from src.shared.helpers.errors.domain_errors import EntityError
 
 
 class DeleteUserUsecase:
@@ -7,6 +8,10 @@ class DeleteUserUsecase:
         self.repo = repo
 
     def __call__(self, idUser: int) -> User:
+
+
+        if type(idUser) != int:
+            raise EntityError("idUser")
 
         user = self.repo.delete_user(idUser)
